feat: home page, newsletter api + singup
We can now easily signup for newsletter.
It is saved in db, with status 0 (which allows for devs to make confirmation email which will set to status 1).
This allows to easily manage newsletter emails.

Whats more we added home page with temporary data like newest projects (will be swapped with API call soon).