improve: remove access from jwt, add admin panel btn
Adds button to admin panel, shown only when access == Admin.

Removes 'access' from JWT, we dont need it in there, instead we will fetch it from DB to get more reactive design e.g. when we change access for an user and he is still logged in, it could be unsafe if he retained permissions in the token.