feat: Submit projects (from user profile)
It is now possible to "Submit New Project". The functionality of that is located in User's profile.

This means we, as a user, can both submit projects and browse the one we have previously submitted.

To existing submitProject() function, we had to add missing 'cost'.

We also added formatCost() available in every js (cuz its in common.js) to more easily distinguish costs.